# openapi.yaml
openapi: 3.0.0
info:
  title: Yachiyo Service Cloud API
  description: |
    YachiyoServiceCloud API 文档
    
    ## 认证说明
    - 所有非白名单接口需要在 Header 中携带 `Authorization: Bearer <token>`
    - Token 过期时间：1小时
    - Token 需通过 `/api/v1/auth/login` 或 `/api/v1/auth/login-by-email` 接口获取
    
    ## 统一响应格式
    ```json
    {
      "code": "200",
      "message": "success", 
      "data": {},
      "detail": null
    }
    ```
  version: 1.0.0
  contact:
    name: Yachiyo Team
servers:
  - url: http://localhost:8881
    description: 本地网关服务
...
paths:
  /api/v1/auth/login:
    post:
      tags: [User Service]
      summary: 用户名密码登录
      description: 使用用户名和密码进行登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
              example:
                code: "200"
                message: success
                data: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                detail: null
  /api/v1/auth/register:
    post:
      tags: [User Service]
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
  /api/v1/auth/send-code:
    post:
      tags: [User Service]
      summary: 发送验证码
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
          description: 邮箱地址
      responses:
        '200':
          description: 发送成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultBoolean'
  /api/v1/auth/login-by-email:
    post:
      tags: [User Service]
      summary: 邮箱验证码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MailLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
  /api/v1/auth/change-password:
    post:
      tags: [User Service]
      summary: 修改密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultBoolean'
  /api/v1/auth/refresh-token:
    post:
      tags: [User Service]
      summary: 刷新 Token
      parameters:
        - name: refreshToken
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
  /api/v1/auth/logout:
    post:
      tags: [User Service]
      summary: 登出
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultBoolean'
  /api/v1/auth/ws-token:
    get:
      tags: [User Service]
      summary: 获取 WebSocket Token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'
  /api/v1/auth/freeze:
    post:
      tags: [User Service]
      summary: 冻结账户
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultBoolean'
...
  /api/v2/column/search:
    get:
      tags: [Content Service]
      summary: 搜索专栏
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: 成功
  /api/v2/column/interaction:
    get:
      tags: [Content Service]
      summary: 获取专栏互动状态
      parameters:
        - name: columnId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
    put:
      tags: [Content Service]
      summary: 专栏互动
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionRequest'
      responses:
        '200':
          description: 成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 格式：Bearer <token>
  schemas:
...
    ResultString:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: string
          nullable: true
        detail:
          type: string
          nullable: true
    ResultBoolean:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: boolean
          nullable: true
        detail:
          type: string
          nullable: true
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          description: 密码
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - email
        - code
      properties:
        username:
          type: string
          description: 用户名（含敏感词过滤）
        password:
          type: string
          description: 密码
        email:
          type: string
          description: 邮箱
        code:
          type: string
          description: 验证码
    MailLoginRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
        code:
          type: string
...
    InteractionRequest:
      type: object
      properties:
        targetId:
          type: integer
          format: int64
        type:
          type: string
          description: 互动类型
        action:
          type: string
          description: 动作
...
    SearchRequest:
      type: object
      properties:
        keyword:
          type: string
        pageNum:
          type: integer
        pageSize:
          type: integer
...